moved the keyring functions back into warg_client
