+++ crates/credentials/Cargo.toml
@@ -12,9 +12,4 @@ repository = { workspace = true }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = { workspace = true }
-indexmap = { workspace = true }
-keyring = { workspace = true }
-secrecy = { workspace = true }
 warg-client = { workspace = true }
-warg-crypto = { workspace = true }
